enable to load chapters
